pages/requirements.txt
streamlit_js_eval==0.1.5
gdown==4.7.1
streamlit_extras==0.2.7
st_click_detector==0.1.3
streamlit==1.30.0
ffmpeg==1.4
django_environ==0.10.0
environ==1.0
Pillow==10.2.0
pydub==0.25.1
Requests==2.31.0
st_pages==0.4.5
streamlit_authenticator
pandas
pydub
playsound
ffmpeg-downloader
ffprobe
